Add Pull Request highlights section #3
If pull requests are large and touch many files, it is sometimes hard to figure out the actual logic changes from refactorings if you're not intimately familiar with the code. In such cases it would be beneficial to add/require that a has PR a highlights section guiding the reviewer's focus.